use inheritance for similar logic
