+++ lib/store_model/types.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "store_model/types/base_single_type"
+require "store_model/types/base_array_type"
 require "store_model/types/json_type"
 require "store_model/types/array_type"
 require "store_model/types/enum_type"
